Use upload artifact Action to create zips
This will create zip files that can be uploaded to the main MIT
production view.

Testing in other views would require renaming a folder manually and
re-zipping. If that becomes a normal process, we can likely use a
manual workflow trigger that prompts for the target view name and
creates the appropriate zip file, but for now that is out of scope.

https://mitlibraries.atlassian.net/browse/IMP-1802